add password requirements to all forms
pre-fill email fofr login screen
